+++ conf/rulesets/solhint-all.js
@@ -6,7 +6,7 @@
 module.exports = Object.freeze({
   rules: {
     'code-complexity': ['warn', 7],
-    'custom-errors': 'warn',
+    'custom-errors': 'error',
     'explicit-types': ['warn'],
     'function-max-lines': ['warn', 50],
     'max-line-length': ['error', 120],