+++ test/rules/best-practises/custom-errors.js
@@ -4,6 +4,7 @@ const {
   assertErrorMessage,
   assertErrorCount,
 } = require('../../common/asserts')
+const { configGetter } = require('../../../lib/config/config-file')
 const linter = require('../../../lib/index')
 const { funcWith } = require('../../common/contract-builder')
 
@@ -71,17 +72,17 @@ describe('Linter - custom-errors', () => {
     assertNoErrors(report)
   })
 
-  it('should not be included in [recommended] config', () => {
+  it('should be included in [recommended] config', () => {
     const code = funcWith(`require(!has(role, account), "Roles: account already has role");
         revert("RevertMessage");
         revert CustomError();
     `)
     const report = linter.processStr(code, {
-      extends: 'solhint:recommended',
-      rules: { 'compiler-version': 'off' },
+      rules: { ...configGetter('solhint:recommended').rules, 'compiler-version': 'off' },
     })
 
-    assertNoErrors(report)
-    assertNoWarnings(report)
+    assertErrorCount(report, 2)
+    assertErrorMessage(report, 0, 'Use Custom Errors instead of require statements')
+    assertErrorMessage(report, 1, 'Use Custom Errors instead of revert statements')
   })
 })