+++ test/rules/naming/named-return-values.js
@@ -1,4 +1,5 @@
 const assert = require('assert')
+const { configGetter } = require('../../../lib/config/config-file')
 const linter = require('../../../lib/index')
 const contractWith = require('../../common/contract-builder').contractWith
 const { assertErrorCount, assertNoErrors, assertWarnsCount } = require('../../common/asserts')
@@ -50,17 +51,15 @@ describe('Linter - named-return-values', () => {
     assert.equal(report.reports[1].message, `4-th return value does not have a name`)
   })
 
-  it('should NOT raise error for solhint:recommended setup', () => {
+  it('should raise warning for solhint:recommended setup', () => {
     const code = contractWith(
-      `function getBalanceFromTokens(address wallet) public returns(address, address, uint256, uint256) { balance = 1; }`
+      `function getBalanceFromTokens(address) public returns(uint256) { balance = 1; }`
     )
 
     const report = linter.processStr(code, {
-      extends: 'solhint:recommended',
-      rules: { 'compiler-version': 'off' },
+      rules: { ...configGetter('solhint:recommended').rules, 'compiler-version': 'off' },
     })
-
-    assertNoErrors(report)
+    assertWarnsCount(report, 1)
   })
 
   it('should raise error for solhint:all setup', () => {